refactor(shared): one PrdDocument type and named string normalisers (§13.3)
`PrdDocument` was declared twice, structurally identically: once in
`server/structuredOutput/types.ts` with its epics and stories spelled out
inline, and once in `src/lib/prdDocument.ts` with them named. Display,
approval and the phase writer could therefore drift about what a valid PRD
is. The type now lives in `shared/prdDocument.ts` and both sides re-export
it, so every existing import site is unchanged.

Five modules also hand-wrote `toStringArray`, and — this is the reason the
shared version is three functions rather than one — they did not agree:

- `src/lib/executionSetupPlan.ts` kept empty strings;
- `src/lib/interviewDocument.ts` and
  `server/phases/executionSetup/runtimeProfile.ts` dropped blank entries but
  returned the survivors with their whitespace intact;
- `src/lib/prdDocument.ts` and `server/phases/interview/sessionSerializer.ts`
  trimmed every entry first, then dropped the ones that became empty.

Collapsing them into one helper would have silently changed what two of the
five callers produce. `shared/stringNormalization.ts` therefore exports
`toStringEntries`, `toNonBlankStringEntries` and `toTrimmedStringEntries`,
each caller imports the one matching what it did before, and
`shared/__tests__/stringNormalization.test.ts` pins the exact inputs on
which the three disagree so a later attempt to merge them fails there.
`toStringValue` and `slugify` were byte-identical in both copies and move
across unchanged. `toStringArray` in `server/structuredOutput/yamlUtils.ts`
is left alone: it coerces dates, numbers and nested objects, which is a
parser concern, not a normaliser.

Departure from the plan: §13.3 also asks for `normalizeOption` and
`normalizeAnswerType` to be shared between the client and server interview
modules. They are not duplicates. The server pair is a parser — wider alias
set, warning collection, and it *throws* on an unsupported `answer_type`;
the client pair is a display normaliser over already-validated output that
must never throw. Sharing them would have merged two different contracts, so
they stay separate with a comment saying why.

Impact: type-only for the PRD change, so no runtime behaviour moves. The
normaliser change is a no-op by construction — every call site kept its
previous semantics under a new name.
